fix(acp): close type-safety gaps from ACP commit (trellis-check)
trellis-check 复核 ACP runtime adapter 发现 3 个类型 bug（均 f7d401a6
引入，对照父 commit 确认），type-only 修复无运行时行为变化：

- acp_process_runtime.py probe_connectivity: 声明返回 int 实际返回
  round(...,1) 的 float；签名改 tuple[bool, str, float]，三处 early
  return (0) 与 api.py:7114 JSON 序列化均兼容 float
- codex_process_manager.py RuntimeCatalogService(self._codex_store):
  RuntimeCodexStore 类型不符 RuntimeCatalogStore 要求；cast 处理
  （匹配 conductor_main_loop.py:1014 既有模式），加 RuntimeCatalogStore
  import（无循环依赖）
- codex_process_manager.py 三路 if/elif 赋 runtime 变量 mypy 窄化报错；
  显式 union 标注 CodexAppServerRuntime | ClaudeProcessRuntime |
  AcpProcessRuntime（BaseProcessRuntime 太窄，不含 write_input_async）

验证: ruff 绿；mypy 3 文件 0 issues（修前 4 errors）；
pytest test_acp_runtime 45 passed + Codex/Claude/catalog 回归 74 passed；
前端 tsc --noEmit 绿、lint 绿。

+++ backend/app/application/acp_process_runtime.py
@@ -130,7 +130,7 @@ class AcpProcessRuntime(BaseProcessRuntime):
 
     async def probe_connectivity(
         self, executor_id: str
-    ) -> tuple[bool, str, int]:
+    ) -> tuple[bool, str, float]:
         """Run a real ACP ``initialize`` handshake against the configured agent.
 
         Spawns the executor's ``command args``, negotiates protocol version 1,
